fix: change border on hover
When we decreased the size of cards, we made the border less than 3px wide but that means we can't have a double border (since that requires at least 3px). Our fix is to increase the size of the border, meaning the border is thicker proportionally than it was before the resize.